requirements.txt
Django==4.2.7
djangorestframework==3.14.0
Pillow==10.1.0
django-crispy-forms==2.1
crispy-bootstrap5
django-filter==23.3
django-leaflet==0.29.0
django-geojson==4.0.0
python-decouple==3.8
gunicorn==21.2.0
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-storages==1.14.2
boto3==1.29.7